+++ tests/PHPUnit/System/TrackerResponseTest.php
@@ -125,7 +125,7 @@ class TrackerResponseTest extends SystemTestCase
         $response = Http::sendHttpRequest($url, 10, null, null, 0, false, false, true);
         $this->assertEquals(200, $response['status']);
 
-        $expected = "This resource is part of Matomo. Keep full control of your data with the leading free and open source <a href='https://matomo.org' target='_blank' rel='noopener noreferrer'>digital analytics platform</a> for web and mobile.";
+        $expected = "This resource is part of Matomo. Keep full control of your data with the leading free and open source <a href='https://matomo.org' target='_blank' rel='noopener noreferrer nofollow'>web analytics & conversion optimisation platform</a>.";
         $this->assertEquals($expected, $response['data']);
     }
 
@@ -136,7 +136,7 @@ class TrackerResponseTest extends SystemTestCase
         $this->assertEquals(400, $response['status']);
 
         $response = $this->sendHttpRequest($url);
-        $expected = "This resource is part of Matomo. Keep full control of your data with the leading free and open source <a href='https://matomo.org' target='_blank' rel='noopener noreferrer'>digital analytics platform</a> for web and mobile.";
+        $expected = "This resource is part of Matomo. Keep full control of your data with the leading free and open source <a href='https://matomo.org' target='_blank' rel='noopener noreferrer nofollow'>web analytics & conversion optimisation platform</a>.";
         $this->assertEquals($expected, $response['data']);
     }